Use junit5 beforeEach
